requirement.txt
Flask
Flask-PyMongo
dnspython
python-dotenv
 python app.py
 from flask import Flask, render_template, request, redirect, url_for
from flask_pymongo import PyMongo
from bson.objectid import ObjectId
from dotenv import load_dotenv
import os
